[SCRUM-207] TEST: CourseService.findCourses의 테스트 코드 구현 (#58)
추천코스 전체조회 API의 테스트 코드를 구현했습니다.
테스트 케이스는 다음과 같습니다.
- 코스 전체 조회 성공 (회원, 필터링)
- 코스 전체 조회 성공 (비회원, 필터링)
- 코스 전체 조회 성공 (조회 결과 없음)
- 코스 전체 조회 실패 (지역 필터링인데 Area 파라미터가 null)
- 코스 전체 조회 실패 (테마 필터링인데 Theme 파라미터가 null)

+++ src/main/java/com/server/running_handai/domain/course/dto/CourseInfoWithDetailsDto.java
@@ -7,7 +7,7 @@ public record CourseInfoWithDetailsDto(
         String thumbnailUrl,
         int distance,
         int duration,
-        int maxElevation,
+        double maxElevation,
         double distanceFromUser,
         int bookmarks,
         boolean isBookmarked,
